Add versions to all tools.
All Galaxy tools should define versions - this is important in an abstract sense for reproducibility and provenance, but in a very practical sense this allows Galaxy to handle tool upgrades more pleasantly. If the tool is upgraded users will be shown a message indicating the tool has changed and rerun or workflows may be altered. Additionally, the most recent version of Galaxy allows multiple versions of the same tool to exist in a single toolbox.