Update all Matlab tools to be relocatable.
The previous way matlab was changing directories made unnecessary assumptions both about how Galaxy was configured (that the files were stored in $GALAXY_ROOT/database/files - a default that is frequently changed by deployers) and where tools were places (under $GALAXY_ROOT/tools - again this should not be forced on deployers and if these tools are ever installable via the tool shed it won't install them here).

The new idiom uses the variable $__tool_directory__ when available to the tool. Unfortunately this was not introduced until the March 2015 release of Galaxy (a rather new release, certainly newer than the Galaxy-M VM), so it fallbacks to the previous default on older Galaxy instances.