Add citations to XCMS tool.
These citations will make it easier for your users to cite XCMS when they use this tools. A link and BibTeX entry will appear at the bottom of the tool form - and users can extract all the citations from a history to summarize a complete analysis.

Added all citations indicated by R's citation("xcms") - if only a subset of those papers are relevant the citations should be pruned accordingly.